Remove redundant imports and move make_square_with_bb to appropriate script.
