+++ core/data/fetch.py
@@ -82,14 +82,6 @@ def img_path_from_row(row: pd.Series, index: int, column="identifier", extra=Non
     return out + f".{extension}"
 
 
-def make_square_with_bb(im, min_size=256, fill_color=(0, 0, 0, 0)):
-    x, y = im.size
-    size = max(min_size, x, y)
-    new_im = Image.new('RGB', (size, size), fill_color)
-    new_im.paste(im, (int((size - x) / 2), int((size - y) / 2)))
-    return new_im
-
-
 @timing
 def fetch_images(df: pd.DataFrame, col: str):
     """
